Refactor logInvalidStartTimes
